Rewrite publication_relation/util_test to separated tests
