--- test_tools/data/maps/publication_relation_util_test/map_1.json
@@ -1,62 +1,0 @@
-{
-  "abstract": "World places and boundaries abstract",
-  "title": "World places and boundaries",
-  "extent": [
-    "-35.0",
-    "-48.5",
-    "179",
-    "81.5"
-  ],
-  "projection": "epsg:3857",
-  "layers": [
-    {
-      "metadata": {},
-      "visibility": true,
-      "opacity": 1,
-      "title": "Defini\u010dn\u00ed body administrativn\u00edch celk\u016f",
-      "className": "HSLayers.Layer.WMS",
-      "singleTile": true,
-      "wmsMaxScale": 0,
-      "legends": [
-        "https%3A%2F%2Fgeoportal.kraj-lbc.cz%2Fcgi-bin%2Fmapserv%3Fmap%3D%2Fdata%2Fgis%2FMapServer%2Fprojects%2Fwms%2Fatlas%2Fadministrativni_cleneni.map%26version%3D1.3.0%26service%3DWMS%26request%3DGetLegendGraphic%26sld_version%3D1.1.0%26layer%3Ddefinicni_body_administrativnich_celku%26format%3Dimage%2Fpng%26STYLE%3Ddefault"
-      ],
-      "maxResolution": null,
-      "minResolution": 0,
-      "url": "https%3A%2F%2Fgeoportal.kraj-lbc.cz%2Fcgi-bin%2Fmapserv%3Fmap%3D%2Fdata%2Fgis%2FMapServer%2Fprojects%2Fwms%2Fatlas%2Fadministrativni_cleneni.map%26",
-      "params": {
-        "LAYERS": "definicni_body_administrativnich_celku",
-        "INFO_FORMAT": "application\/vnd.ogc.gml",
-        "FORMAT": "image\/png",
-        "FROMCRS": "EPSG:3857",
-        "VERSION": "1.3.0"
-      },
-      "dimensions": {}
-    },
-    {
-      "metadata": {},
-      "visibility": true,
-      "opacity": 1,
-      "title": "Hranice",
-      "className": "HSLayers.Layer.WMS",
-      "singleTile": true,
-      "url": "http://localhost:8000/geoserver/testuser1_wms/ows",
-      "params": {
-        "LAYERS": "hranice",
-        "FORMAT": "image\/png"
-      }
-    },
-    {
-      "metadata": {},
-      "visibility": true,
-      "opacity": 1,
-      "title": "Mista",
-      "className": "HSLayers.Layer.WMS",
-      "singleTile": true,
-      "url": "http://localhost:8000/geoserver/testuser1_wms/ows",
-      "params": {
-        "LAYERS": "mista",
-        "FORMAT": "image\/png"
-      }
-    }
-  ]
-}
